Move navigation UI code to helper class
